+++ addons/overrides/data/config_vehicles.hpp
@@ -1,3 +1,0 @@
-#include "wm_patches\vehicles\tie_ln\config_vehicles.hpp"
-#include "wm_patches\vehicles\tie_bomber\config_vehicles.hpp"
-